test: extend cypress common commands
