+++ solution/ui/e2e/cypress/e2e/homepage.spec.cy.js
@@ -139,9 +139,7 @@ describe("Homepage", { scrollBehavior: "center" }, () => {
     it("allows a user to go back to the homepage by clicking the top left link", () => {
         cy.viewport("macbook-15");
         cy.visit("/42/430/");
-        cy.contains("Medicaid & CHIP eRegulations").click();
-
-        cy.url().should("eq", Cypress.config().baseUrl + "/");
+        cy.goHome();
     });
 
     it("has the correct descriptive text", () => {
