+++ solution/ui/e2e/cypress/support/e2e.js
@@ -28,13 +28,15 @@ import "cypress-axe";
 import "cypress-plugin-tab";
 
 import { getResources, getSearchGovResources } from "./api-request-commands";
-import { checkBlockingModal, checkFlashBanner, jumpToRegulationPart, jumpToRegulationPartSection } from "./common-commands";
+import { checkBlockingModal, checkFlashBanner, goHome } from"./common-commands/header";
+import { jumpToRegulationPart, jumpToRegulationPartSection } from "./common-commands/jumpTo";
 import { validateSchema } from "./validate-schema-command";
 
 Cypress.Commands.add("checkBlockingModal", checkBlockingModal);
 Cypress.Commands.add("checkFlashBanner", checkFlashBanner);
 Cypress.Commands.add("getResources", getResources);
 Cypress.Commands.add("getSearchGovResources", getSearchGovResources);
+Cypress.Commands.add("goHome", goHome);
 Cypress.Commands.add("jumpToRegulationPart", jumpToRegulationPart);
 Cypress.Commands.add("jumpToRegulationPartSection", jumpToRegulationPartSection);
 Cypress.Commands.add("validateSchema", validateSchema);
